+++ Common/src/main/java/fuzs/airhop/mixin/client/accessor/LivingEntityAccessor.java
@@ -1,4 +1,4 @@
-package fuzs.airhop.mixin.accessor;
+package fuzs.airhop.mixin.client.accessor;
 
 import net.minecraft.world.entity.LivingEntity;
 import org.spongepowered.asm.mixin.Mixin;
@@ -6,6 +6,7 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(LivingEntity.class)
 public interface LivingEntityAccessor {
+
     @Accessor
     boolean getJumping();