+++ Forge/src/main/java/fuzs/airhop/AirHopForge.java
@@ -2,16 +2,13 @@ package fuzs.airhop;
 
 import fuzs.airhop.capability.AirHopsCapability;
 import fuzs.airhop.handler.PlayerFallHandler;
-import fuzs.airhop.handler.PlayerSyncHandler;
 import fuzs.airhop.init.ModRegistry;
 import fuzs.puzzleslib.capability.ForgeCapabilityController;
 import fuzs.puzzleslib.core.CoreServices;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.common.capabilities.CapabilityToken;
 import net.minecraftforge.event.TickEvent;
-import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.event.entity.living.LivingFallEvent;
-import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent;
@@ -40,16 +37,6 @@ public class AirHopForge {
         });
         MinecraftForge.EVENT_BUS.addListener((final LivingFallEvent evt) -> {
             evt.setDistance(playerFallHandler.onLivingFall(evt.getEntity(), evt.getDistance(), evt.getDamageMultiplier()));
-        });
-        final PlayerSyncHandler playerSyncHandler = new PlayerSyncHandler();
-        MinecraftForge.EVENT_BUS.addListener((final PlayerEvent.PlayerLoggedInEvent evt) -> {
-            playerSyncHandler.onEntityJoinLevel(evt.getEntity(), evt.getEntity().getLevel());
-        });
-        MinecraftForge.EVENT_BUS.addListener((final EntityJoinLevelEvent evt) -> {
-            playerSyncHandler.onEntityJoinLevel(evt.getEntity(), evt.getLevel());
-        });
-        MinecraftForge.EVENT_BUS.addListener((final PlayerEvent.PlayerChangedDimensionEvent evt) -> {
-            playerSyncHandler.onEntityJoinLevel(evt.getEntity(), evt.getEntity().level);
         });
     }
 }